chore(docs): updating docs to match new recursion interfacee (#4187)
# Description

This PR documents the new recursion interface.

## Problem\*

Resolves #4139 
Resolves AztecProtocol/dev-rel#148

## Summary\*

- Removes the usage of `aggregation object` while retaining the mention
that recursive proofs are actually aggregation of proofs
- Removes the mention to add the size of public inputs to the proof, as
they are now separate
- Removes some leftover explainer in the `verify_proof` reference

## Additional Context

`noir-examples` is actually not working with this new interface, as I'm
blocked by a weird issue where proofs give me my dear friend
`unreachable`

Will debug that next week with @TomAFrench but I think we're good to
push the docs for now.

---------

Co-authored-by: Cat McGee <helloworld@mcgee.cat>