feat: remove replacement of boolean range opcodes with `AssertZero` opcodes (#4107)
…

# Description

## Problem\*

Complement to https://github.com/AztecProtocol/aztec-packages/pull/4164.
I'm going to leave this in draft until we've pushed out a release of
`bb` containing the optimisation.

This optimisation is being moved to be a responsibility of the backend
so we shouldn't do it here.

## Summary\*



## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.