feat: remove predicate from `sort` intrinsic function (#4228)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR removes the predicate from `AcirContext.sort()`. I'm not sure
why this exists as we should always be able to perform a sorting on an
array of values. We also list `Intrinsic::Sort` as having no side
effects so it doesn't make sense for us to be using the side effects
predicate here.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.