chore: automatically recompile `noir_js` test case programs (#4236)
# Description

## Problem\*

[CI is currently failing in
master](https://github.com/noir-lang/noir/actions/runs/7755136791/job/21150220295)
due to the new testcase added in #4101 not being added to the docker
test flow.

## Summary\*

This PR adds a compilation step to `yarn test` which recompiles all of
the test programs with the currently installed version of nargo. This is
similar to what we're doing in `noir_codegen`.


## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.