fix: correct result when assigning shared arrays in unconstrained code (#4210)
# Description

## Problem\*

Resolves #3795

## Summary\*

`Instruction::IncrementRc` in Brillig actually only increments the
reference count on a copy of an array's metadata if the array was loaded
from a reference (while the loaded array is a reference rather than a
copy). This is suboptimal since it was meant to be shared across
references to the array. This is what caused the original issue where
mutating the reference saw a different reference count and thought it
was safe to mutate without copying first.

I've added a somewhat hacky check in the method to increment reference
counts to check if the array was loaded from a reference. If so, we load
a fresh value, increment rc on that, and re-store it to update the
metadata in the original array.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Maxim Vezenov <mvezenov@gmail.com>