chore: Brillig call return type check (#4282)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

Followup to #4280 which added a Brillig call check in SSA and
https://github.com/noir-lang/noir/pull/4281 which type checks the
arguments to an unconstrained call.

## Summary\*

#4280 added a type check that is only caught during SSA as it was
simpler for comparing the runtimes of function as well as encompassing
any use of generics for an unconstrained function definition. However,
we can still catch the error earlier in type check.

I also removed the specific tag-line as mentioned in this
[comment](https://github.com/noir-lang/noir/pull/4280#discussion_r1480468808).
However, as `Vec` is the only non-obvious place a slice may be used I
think the specific tag-line is ok in this case.

## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.