chore(docs): codegen reference docs for `nargo` CLI (#4149)
# Description

## Problem\*

Resolves #4136 
Resolves https://github.com/AztecProtocol/dev-rel/issues/128
## Summary\*

This PR codegens the reference material for the nargo CLI to avoid us
having to copy documentation from the code over to a markdown file.

The exact format of the codegenned markdown isn't final and we can
iterate on this but it guarantees completeness.

## Additional Context


## Documentation\*

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.