fix: SSA typing for right shifts (#4302)
# Description

## Problem\*

Partial work towards https://github.com/noir-lang/noir/issues/4275

## Summary\*

Previously, SSA was issuing (lhs_type, Field) DIV for right shifts.
However this is wrong if we support shifts by the bit size (`u1 >> 1`)

This is interpreted in acir_gen as an euclidean division with bit_size =
lhs_type.bit_size(). However, u1 >> 1 worked because of a + 1 that is in
euclidean division
https://github.com/noir-lang/noir/blob/0e073037b00212fd17fc8ca9c531b567614eb4c5/compiler/noirc_evaluator/src/ssa/acir_gen/acir_ir/acir_variable.rs#L703
(and it failed in unconstrained functions since they don't codegen that
extraneous + 1)

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: TomAFrench <tom@tomfren.ch>