chore(docs): Add error info about libc++ and contextual search (#4328)
# Description

- Adds contextual search to the docs (so it will always search within
the selected version).
- Adds a note about installing libc++-dev when getting the associated
backend error. I added to the migration notes page instead of the
installation page because people will hit this when they run `nargo
prove`. There are other messages about backend errors on that page, so
it seemed like a more appropriate place.

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>
Co-authored-by: Savio <72797635+Savio-Sou@users.noreply.github.com>