fix(noir-mirror): don't update .gitrepo on push  (#4555)
We no longer need a commit on aztec side, and if we forgo this we're a
bit more robust actually because then the base URL is always the last
pull
There is a heuristic for finding the last commit that should be
considered 'synced' for commit message computation purposes. The
heuristic failed because
1. there were extra commits in aztec packages after our pull
2. then when we fell back to our gitrepo commit, it was changed on sync

To fix this, we can exploit not needing to actually know our last pushed
commit as we know squash. This way, our 'commit' variable in our
.gitrepo is always our last *pull* and not pull or push