fix(acir): Array dynamic flatten (#4351)
# Description

## Problem\*

Resolves #4258 

## Summary\*

There are some instances where we want to flatten a single AcirValue
into its AcirVars. We were operating under the assumption that we cannot
flatten dynamic arrays. I have changed the `flatten` method to be on an
`AcirContext` so that we can read from dynamic memory when appropriate.

I have added two tests performing what is done in the issue.
1. Returning a dynamic array from main
2. Preparing a dynamic array as inputs to a black box function

## Additional Context

Further investigation should be done to see if we should fully remove
the `flatten()` method on `AcirValue` and move to only using the
`flatten` which lives on the AcirContext. I chose to leave that for
separate work though and scope this PR to just the bug fix in the issue.

## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <tom@tomfren.ch>