feat: DAP Preflight and debugger compilation options (#4185)
# Description

## Problem\*

Part of #3015 

## Summary\*

This PR adds a preflight mode to DAP in order to make it easier to
identify and report back problems to the user when compiling the project
for debugging. This preflight mode is invoked from the VS.Code extension
before starting the debugging session, and with the same arguments as
those that will be used for the session. If the compiler finds any error
either loading or compiling the project, the error is reported to stderr
which allows the extension to parse the output and present the
diagnostic messages to the user.

This also changes the default compilation mode to output Brillig code
and adds new commands line options to Nargo's `debug` command and launch
options to the DAP mode to control the mode and whether to inject
instrumentation code to track variables or not.

The `debug` options are:
- `--acir-mode`, force output of ACIR, which disables instrumentation by
default
- `--skip-instrumentation={true,false}` to control injection of
instrumentation code to track variables values during the debugging
session.

Similarly, for DAP two launch options can be provided: `generateAcir`
and `skipInstrumentation`.

## Additional Context

The default is to output in Brillig mode with instrumentation for
tracking variables, as this makes it easier to follow along with
stepping through the code. If ACIR mode is selected, instrumentation is
disabled by default. Instrumentation can be forcefully enabled or
disabled by the provided CLI option.

## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [X] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Martin Verzilli <martin.verzilli@gmail.com>