chore: Test for printing array of strings (#4389)
# Description

## Problem\*

Resolves #2903 

## Summary\*

I added a function `regression_2903` to the `debug_logs` test that shows
we can now accurately print arrays of strings. This was fixed in
previous formatting work. I also removed `std::` prefixes from `println`
now that we have a prelude.

## Additional Context


## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.