feat: update error message when trying to load workspace as dependency (#4393)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This addresses an issue that a user in the discord is running into where
they're trying to load a workspace as a dependency and can't figure out
how to fix their Nargo.toml.

## Additional Context

We should probably be embedding links to the docs but that would require
some setup so that we point to the correct version, etc.


## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

Co-authored-by: kevaundray <kevtheappdev@gmail.com>