chore: Add #[recursive] Explainer to Documentation (#4399)
# Description

## Problem

Step towards [#4392](https://github.com/noir-lang/noir/issues/4392)

The documentation for Recursive Proofs was lacking information on the
newly introduced `#[recursive]` attribute, which is crucial for
understanding how to mark circuits for recursive proof generation.

## Summary

This pull request updates the Recursive Proofs documentation page to
include a comprehensive section on the `#[recursive]` attribute. It
explains the attribute's purpose, how it should be used, and provides an
example demonstrating its application within a circuit definition.

## Additional Context

The introduction of the `#[recursive]` attribute simplifies the process
of designating circuits for recursive proofs, eliminating the need for
manual flagging in the tooling infrastructure.

## Documentation

- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Savio <72797635+Savio-Sou@users.noreply.github.com>