chore(ssa): Remove mem2reg run before flattening (#4415)
# Description

## Problem\*

Before https://github.com/noir-lang/noir/pull/4240 we needed mem2reg to
be run as to not panic when fetching slice lengths.

## Summary\*

After the linked PR we have an improved startegy for tracking slice
capacities by generating a slice capacities map before merging of
values. This should enable us to remove a mem2reg pass that is run
before flattening.

## Additional Context


## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.