chore: do not panic when dividing by zero (#4424)
# Description

## Problem\*

Resolves #2480

## Summary\*
In BrilligVM, we have now the same behaviour regarding division by 0.
Whether it is a field or integer division, we return 0 when dividing by
zero. Since we have a constraint which checks that the inverse times
itself is 1, this constraint will always fail if we divide by zero
(instead of panic or returning an error).


## Additional Context



## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.