chore(ci): enforce formatting of noir code in CI (#4422)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

We currently format everything in the repository except our noir source
code. If we enforce this internally then we'll uncover issues in the
formatter earlier and provide a good example of what Noir source should
look like.

We then now run `nargo fmt --check` on the stdlib and
`test_programs/execution_success`

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.